requirements.txt
fastapi==0.111.2
uvicorn[standard]==0.23.2
numpy==1.26.4
pandas==2.1.1
python-dotenv==1.0.1
openai==1.7.3
faiss-cpu==1.7.4
rank_bm25==0.2.2
jinja2==3.1.2
openai==1.7.3
python-multipart==0.0.20
